stitch clean up
