+++ client/public/index.html
@@ -49,7 +49,7 @@
       To begin the development, run `npm start` or `yarn start`.
       To create a production bundle, use `npm run build` or `yarn build`.
     -->
-    <script src="/bundle.js"></script>
+    <!-- <script src="/bundle.js"></script> -->
 </body>
 
 </html>
