+++ client/src/App.js
@@ -9,7 +9,7 @@ import Layout from "./components/Layout"
 import Subscription from "./components/Subscription"
 
 import FileInput from "./components/FileInput"
-import Artist from "./pages/Artist"
+import AWS from "./stitch/app"
 import Admin from "./pages/Admin"
 import {StoreProvider} from "./utils/globalContext"
 import Login from "./components/LoginForm"
@@ -30,7 +30,7 @@ function App() {
         <Route exact path='/activate/:token' component={ActivateUser}/>
         <Route exact path='/listener' component={Center}/>
         <Route exact path='/admin/dashboard' component={Admin}/>
-        <Route exact path='/test' component={Artist}/>
+        <Route exact path='/test' component={AWS}/>
         <Route exact path='/subscription' component={Subscription}/>
         <Route exact path='/artistpage' component={ArtistPage}/>
