+++ client/src/stitch/app.js
@@ -1,11 +1,10 @@
 import React from "react";
-import axios from "axios";
 import FileInput from "../components/FileInput";
-import { handleSignup, handleLogin } from "../utils/stitch";
+import { handleSignup, handleLogin } from "./authentication";
 import { useStoreContext } from "../utils/globalContext";
 import { AwsServiceClient, AwsRequest } from "mongodb-stitch-browser-services-aws";
 import { RemoteMongoClient } from "mongodb-stitch-browser-services-mongodb-remote";
-import { stitchClient } from "../utils/stitch";
+import { stitchClient } from "./authentication";
 import songsAPI from "../utils/songsAPI"
 // import { UserContext } from "../utils/UserContext";
 
@@ -29,7 +28,7 @@ const convertAudioToBSONBinaryObject = (file) => {
   });
 };
 
-function ArtistPage() {
+function AWS() {
   const [global, dispatch] = useStoreContext();
   let userEmail = global.user.email;
   let userPass = global.user.password;
@@ -116,4 +115,4 @@ function ArtistPage() {
   );
 }
 
-export default ArtistPage;
+export default AWS;
